refactor: 컴포넌트 및 상수 분리
