+++ apps/client/src/constants/user.ts
@@ -9,3 +9,21 @@ export const OAUTH_BADGE: Record<
     KAKAO: KakaoBadge,
     NAVER: NaverBadge,
 };
+
+export const GENDERS = [
+    {
+        label: 'MALE',
+        value: '남성',
+        checked: false,
+    },
+    {
+        label: 'FEMAIL',
+        value: '여성',
+        checked: false,
+    },
+    {
+        label: 'UNKNOWN',
+        value: '선택하지 않음',
+        checked: false,
+    },
+] as const;
